+++ adapters/Vungle/VungleAdapter/Bidding/GADMediationVungleInterstitial.m
@@ -73,7 +73,8 @@
   if (!self.desiredPlacement.length) {
     NSError *error = GADMAdapterVungleErrorWithCodeAndDescription(
         GADMAdapterVungleErrorInvalidServerParameters,
-        @"Missing or invalid Placement ID configured for this ad source instance in the AdMob or Ad Manager UI.");
+        @"Missing or invalid Placement ID configured for this ad source instance in the AdMob or "
+        @"Ad Manager UI.");
     _adLoadCompletionHandler(nil, error);
     return;
   }