+++ adapters/Vungle/VungleAdapter/Bidding/GADMediationVungleNativeAd.m
@@ -81,7 +81,8 @@
   if (!self.desiredPlacement) {
     NSError *error = GADMAdapterVungleErrorWithCodeAndDescription(
         GADMAdapterVungleErrorInvalidServerParameters,
-        @"Missing or invalid Placement ID configured for this ad source instance in the AdMob or Ad Manager UI.");
+        @"Missing or invalid Placement ID configured for this ad source instance in the AdMob or "
+        @"Ad Manager UI.");
     _adLoadCompletionHandler(nil, error);
     return;
   }